Form Validation

+++ app/Http/Controllers/PharmacyController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Area;
 use App\Models\Pharmacy;
 use Illuminate\Http\Request;
+use App\Http\Requests\PharmacyRequest;
 
 class PharmacyController extends Controller
 {
@@ -14,9 +15,7 @@ class PharmacyController extends Controller
     public function index()
     {
         $allPharmacies = Pharmacy::withTrashed()->paginate(5);
-        return view('pharmacies.index', [
-            'pharmacies' => $allPharmacies,
-        ]);
+        return view('pharmacies.index', ['pharmacies' => $allPharmacies]);
     }
 
     /**
@@ -24,44 +23,24 @@ class PharmacyController extends Controller
      */
     public function create()
     {
-        //
         $areas = Area::all();
-        return view('pharmacies.create',
-            compact('areas'));
+        return view('pharmacies.create', compact('areas'));
     }
 
     /**
      * Store a newly created resource in storage.
      */
-//    public function store(Request $request)
-//    {
-//        //
-//        $allPharmacies = Pharmacy::create($request->all());
-//            return redirect()->route('pharmacies.index')->with('success','Pharmacy created successfully!');
-//    }
-    public function store(Request $request)
+    public function store(PharmacyRequest $request)
     {
-//        // Validate the request data
-//        $request->validate([
-//            'name' => 'required',
-//            'priority' => 'required|integer',
-//            'area' => 'required',
-//            'avatar' => 'nullable|image|max:2048', // max file size is 2MB
-//        ]);
-
         $pharmacy = Pharmacy::create($request->all());
-        // Handle the image upload, if provided
         if ($request->hasFile('avatar')) {
             $avatarName = time().'.'.$request->avatar->extension();
-//            $request->avatar->storeAs('public/images', $avatarName);
             $request->avatar->move('images/pharmacies/', $avatarName);
             $pharmacy->avatar = 'images/pharmacies/' . $avatarName;
         } else {
             $avatarName = null;
         }
-
         $pharmacy->save();
-
         return redirect()->route('pharmacies.index')->with('success', 'Pharmacy created successfully!');
     }
 
@@ -86,23 +65,18 @@ class PharmacyController extends Controller
     /**
      * Update the specified resource in storage.
      */
-//    public function update(Request $request, Pharmacy $pharmacy)
-//    {
-//        $pharmacy->update($request->all());
-//        return redirect()->route('pharmacies.index');
-//    }
-    public function update(Request $request, Pharmacy $pharmacy)
+    public function update(PharmacyRequest $request, Pharmacy $pharmacy)
     {
-            $pharmacy->update($request->except('avatar'));
-            if ($request->hasFile('avatar')) {
-                $old_avatar = $pharmacy->avatar;
-                $avatar = $request->avatar;
-                $avatar_new_name = time().'.'.$request->avatar->extension();
-                if ($avatar->move('images/pharmacies/', $avatar_new_name)) {
-                    unlink($old_avatar);
-                }
-                $pharmacy->avatar = 'images/pharmacies/' . $avatar_new_name;
+        $pharmacy->update($request->except('avatar'));
+        if ($request->hasFile('avatar') && $request->file('avatar')->isValid()) {
+            $old_avatar = $pharmacy->avatar;
+            $avatar = $request->avatar;
+            $avatar_new_name = time().'.'.$request->avatar->extension();
+            if ($avatar->move('images/pharmacies/', $avatar_new_name)) {
+                unlink($old_avatar);
             }
+            $pharmacy->avatar = 'images/pharmacies/' . $avatar_new_name;
+        }
         $pharmacy->save();
         return redirect()->route('pharmacies.index');
     }
